# swagger.yaml
openapi: 3.0.0
info:
  title: 4Events API
  description: "<p>This is an API belonging to the 4events group that is dedicated to putting people who organize directed activities in contact with people who are looking for someone to organize them.</p><p>Here you will find everything you are looking for and you can offer any activity that you want to organize</p>"
  version: 0.1.9
servers:
  - url: https://services.4events.net
    description: Main (production) server
  - url: http://localhost:3000
    description: Developer (Local)
paths:
...
  /api/v1/events:
    get:
      security:
        - {}
        - bearerAuth: []
      summary: List of events. Only returns the event date will be after now.
      
      parameters:
        - in: query
          name: sort
          schema:
            type: string
          required: false
          description: If 'asc', returns the date sorted in ascending order else the order is descendig.
        - in: query
          name: skip
          schema:
            type: integer
          required: false
          description: Skip the events indicated in a pagination
        - in: query
          name: limit
          schema:
            type: integer
          required: false
          description: Maximum events returned in a pagination.
        - in: query
          name: title
          schema:
            type: string
          required: false
          description: Searches by word content and is not case sensitive.
        - in: query
          name: description
          schema:
            type: string
          required: false
          description: Searches by word content and is not case sensitive
        - in: query
          name: indoor
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
          required: false
          description: Searches by event in indoor or outdoor
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          required: false
          description: Searches by event tags
        - in: query
          name: price
          schema:
            type: string
          required: false
          description: Supports single value 258.0 or between two values 0-258.3 Supports decimals with .   
        - in: query
          name: lat
          schema:
            type: number
          required: false
          description: 'Supported format:format Decimal degrees (DD):41.40338, not commas.'
        - in: query
          name: long
          schema:
            type: number
          required: false
          description: 'Supported format:format Decimal degrees (DD):41.40338, not commas.'
        - in: query
          name: distance_m
          schema:
            type: number
          description: distance in meters
          required: false 
        - in: query
          name: userName
          schema:
            type: string
          description: User name. Filter by user events
          required: false   
        - in: header
          name: x-access-token
          schema:
            type: string
          required: false

      tags:
        - GET PUT and DELETE events

      responses:
          '200':    # status code
            description: A JSON object with data
            content:
              application/json:
                schema: 
                  $ref: '#/components/schemas/eventsAuth'
          '403':    
            description: The user does not have privileges for this action
          '422':
            description: Unprocessable Entinty

    post:
      security:
        - bearerAuth: []
      summary: Insert a new event. 
      requestBody:
        content:
          multipart/form-data:
            schema: 
              type: object
              required:
                - title
                - price
                - indoor
                - date
                - duration
                - tags
              properties:
                title:
                  type: string
                  description: Title of event
                description: 
                  type: string
                price: 
                  type: number
                max_places:
                  type: integer
                date:
                  type: string
                  format: date-time
                  description: values of date YYYY/MM/DD HH:MI:SS
                duration:
                  type: integer
                  description: duration value in minutes
                photo:
                  type: string
                  format: binary
                indoor:
                  type: string
                  enum:
                    - 'true'
                    - 'false'
                latitude:
                  type: number
                longitude:
                  type: number
                address:
                  type: string
                city:
                  type: string
                postal_code:
                  type: string
                country:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      parameters:
        - in: header
          name: x-access-token
          schema:
            type: string           
        - in: query
          name: lang
          description: Error return language. Optional value. default value 'en' or header 'Accept-Language'
          schema:
            type: string
            enum:
              - en
              - es

      tags:
      - GET PUT and DELETE events

      responses:
        '201':    # status code
          description: A JSON object with data
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/eventsAuth'
        '403':    
          description: The user does not have privileges for this action
        '422':
          description: Unprocessable Entinty  
    
    put:
      security:
        - bearerAuth: []
      summary: Modify an event by _id. This operation can only be performed by the event owner or a super administrator
      
      parameters:
      - in: path
        name: _id 
        schema:
          type: string
          format: uuid
        required: true
        description: event _id
      
      - in: header
        name: x-access-token
        schema:
          type: string
      - in: query
        name: lang
        description: Error return language. Optional value. default value 'en' or header 'Accept-Language'
        schema:
          type: string
          enum:
            - en
            - es
      requestBody:
        content:
          multipart/form-data:
            schema: 
              type: object
              properties:
                title:
                  type: string
                  description: Title of event
                description: 
                  type: string
                price: 
                  type: number
                max_places:
                  type: integer
                date:
                  type: string
                  format: date-time
                  description: values of date YYYY/MM/DD HH:MI:SS
                duration:
                  type: integer
                  description: duration value in minutes
                photo:
                  type: string
                  format: binary
                indoor:
                  type: string
                  enum:
                    - 'true'
                    - 'false'
                latitude:
                  type: number
                longitude:
                  type: number
                address:
                  type: string
                city:
                  type: string
                postal_code:
                  type: string
                country:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      
      tags:
      - GET PUT and DELETE events

      responses:
        '201':    # status code
          description: A JSON object with data
          content:
              application/json:
                schema: 
                  type: object
                  properties:
                    result:
                      type: string
                      format: uuid
        '403':    
          description: The user does not have privileges for this action
        '422':
          description: Unprocessable Entinty  
    
    delete:
      security:
        - bearerAuth: []
      summary: Delete an event by _id.
      
      parameters:
      - in: path
        name: _id 
        schema:
          type: string
          format: uuid
        required: true
        description: event _id
      
      - in: header
        name: x-access-token
        schema:
          type: string
      
      tags:
      - GET PUT and DELETE events

      responses:
        '201':    # status code
          description: A JSON object with data
          content:
              application/json:
                schema: 
                  type: object
                  properties:
                    result:
                      type: string
                      format: uuid
        '403':    
          description: The user does not have privileges for this action
        '422':
          description: Unprocessable Entinty  
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Geometry:
        type: object
        description: GeoJSon geometry
        
        required:
          - type
        externalDocs:
          url: http://geojson.org/geojson-spec.html#geometry-objects
        properties:
          type:
            type: string
            enum:
            - Point
            - LineString
            - Polygon
            - MultiPoint
            - MultiLineString
            - MultiPolygon
            description: the geometry type
    Point2D:
      type: array
      maxItems: 2
      minItems: 2
      items:
        type: number
    Point:
      type: object
      description: GeoJSon geometry
      externalDocs:
        url: http://geojson.org/geojson-spec.html#id2
      allOf:
        - $ref: "#/components/schemas/Geometry"
        - properties:
            coordinates:
              $ref: '#/components/schemas/Point2D'
...
    eventsAuth:
      type: object
      properties: 
        total:
          type: integer
          description: Total events with filter and without paging 
        events:
          type: array
          items:
            type: object
            properties:
              _id: 
                type: string
                format: uuid
              tags:
                type: array
                items: 
                  type: string
              title:
                type: string
              description:
                type: string
              price:
                type: number
              max_places:
                type: integer
              date:
                type: string
                format: date-time
              duration:
                type: number
              indoor:
                type: boolean
              location:
                type: array
                items:
                  $ref: "#/components/schemas/Point"
              address:
                type: string
              city: 
                type: string
              postal_code:
                type: string
              country:
                type: string
              created_date:
                type: string
                format: date-time
              detailsOwn:
                type: object
                properties:
                  _id: 
                    type: string
                    format: uuid
                  username: 
                    type: string
                    description: Show only if user is authenticated
                  email:
                    type: string
                    format: email
                    description: Show only if user is authenticated
                  nickname: 
                    type: string
                  image:
                    type: string
                    format: uri
              photo:
                type: string
                format: uri
              assistants_count:
                type: integer
              favorite_events_count:
                type: integer
              available_places:
                type: integer
              event_distance:
                type: number
                description: Distance to event in meters.
              isOwner:
                type: boolean
              isFavorite:
                type: boolean
              isAssistant:
                type: boolean
...